MDL-55528 assignfeedback_editpdf: Update to use fileconverter_unoconv

+++ mod/assign/feedback/editpdf/classes/document_services.php
@@ -145,12 +145,13 @@ EOD;
      * This function will search for all files that can be converted
      * and concatinated into a PDF (1.4) - for any submission plugin
      * for this students attempt.
+     *
      * @param int|\assign $assignment
      * @param int $userid
      * @param int $attemptnumber (-1 means latest attempt)
-     * @return array(stored_file)
+     * @return combined_document
      */
-    public static function list_compatible_submission_files_for_attempt($assignment, $userid, $attemptnumber) {
+    protected static function list_compatible_submission_files_for_attempt($assignment, $userid, $attemptnumber) {
         global $USER, $DB;
 
         $assignment = self::get_assignment_from_param($assignment);
@@ -171,10 +172,11 @@ EOD;
 
         // User has not submitted anything yet.
         if (!$submission) {
-            return $files;
+            return new combined_document();
         }
 
         $fs = get_file_storage();
+        $converter = new \core_files\converter();
         // Ask each plugin for it's list of files.
         foreach ($assignment->get_submission_plugins() as $plugin) {
             if ($plugin->is_enabled() && $plugin->is_visible()) {
@@ -183,7 +185,7 @@ EOD;
                     if ($file instanceof \stored_file) {
                         if ($file->get_mimetype() === 'application/pdf') {
                             $files[$filename] = $file;
-                        } else if ($convertedfile = $fs->get_converted_document($file, 'pdf')) {
+                        } else if ($convertedfile = $converter->start_conversion($file, 'pdf')) {
                             $files[$filename] = $convertedfile;
                         }
                     } else {
@@ -199,9 +201,28 @@ EOD;
                         $record->filepath = '/';
                         $record->filename = $plugin->get_type() . '-' . $filename;
 
-                        $htmlfile = $fs->create_file_from_string($record, $file);
-                        $convertedfile = $fs->get_converted_document($htmlfile, 'pdf');
-                        $htmlfile->delete();
+                        $htmlfile = $fs->get_file($record->contextid,
+                                $record->component,
+                                $record->filearea,
+                                $record->itemid,
+                                $record->filepath,
+                                $record->filename);
+
+                        $newhash = sha1($file);
+
+                        // If the file exists, and the content hash doesn't match, remove it.
+                        if ($htmlfile && $newhash !== $htmlfile->get_contenthash()) {
+                            $htmlfile->delete();
+                            $htmlfile = false;
+                        }
+
+                        // If the file doesn't exist, or if it was removed above, create a new one.
+                        if (!$htmlfile) {
+                            $htmlfile = $fs->create_file_from_string($record, $file);
+                        }
+
+                        $convertedfile = $converter->start_conversion($htmlfile, 'pdf');
+
                         if ($convertedfile) {
                             $files[$filename] = $convertedfile;
                         }
@@ -209,18 +230,21 @@ EOD;
                 }
             }
         }
-        return $files;
+        $combineddocument = new combined_document();
+        $combineddocument->set_source_files($files);
+
+        return $combineddocument;
     }
 
     /**
-     * This function return the combined pdf for all valid submission files.
+     * Fetch the current combined document ready for state checking.
+     *
      * @param int|\assign $assignment
      * @param int $userid
      * @param int $attemptnumber (-1 means latest attempt)
-     * @return stored_file
+     * @return combined_document
      */
-    public static function get_combined_pdf_for_attempt($assignment, $userid, $attemptnumber) {
-
+    public static function get_combined_document_for_attempt($assignment, $userid, $attemptnumber) {
         global $USER, $DB;
 
         $assignment = self::get_assignment_from_param($assignment);
@@ -245,107 +269,52 @@ EOD;
         $filename = self::COMBINED_PDF_FILENAME;
         $fs = \get_file_storage();
 
-        if (!$combinedpdf = $fs->get_file($contextid, $component, $filearea, $itemid, $filepath, $filename)) {
-            return self::generate_combined_pdf_for_attempt($assignment, $userid, $attemptnumber);
+        $combinedpdf = $fs->get_file($contextid, $component, $filearea, $itemid, $filepath, $filename);
+        if ($combinedpdf && $submission) {
+            if ($combinedpdf->get_timemodified() < $submission->timemodified) {
+                // The submission has been updated since the PDF was generated.
+                $combinedpdf = false;
+            } else if ($combinedpdf->get_contenthash() == self::BLANK_PDF_HASH) {
+                // The PDF is for a blank page.
+                $combinedpdf = false;
+            }
         }
-        if ($submission && ($combinedpdf->get_timemodified() < $submission->timemodified ||
-                $combinedpdf->get_contenthash() == self::BLANK_PDF_HASH)) {
-            return self::generate_combined_pdf_for_attempt($assignment, $userid, $attemptnumber);
+
+        if (empty($combinedpdf)) {
+            // The combined PDF does not exist yet. Return the list of files to be combined.
+            return self::list_compatible_submission_files_for_attempt($assignment, $userid, $attemptnumber);
+        } else {
+            // The combined PDF aleady exists. Return it in a new combined_document object.
+            $combineddocument = new combined_document();
+            return $combineddocument->set_combined_file($combinedpdf);
         }
-        return $combinedpdf;
     }
 
     /**
-     * This function will take all of the compatible files for a submission
-     * and combine them into one PDF.
+     * This function return the combined pdf for all valid submission files.
+     *
      * @param int|\assign $assignment
      * @param int $userid
      * @param int $attemptnumber (-1 means latest attempt)
-     * @return stored_file
+     * @return combined_document
      */
-    public static function generate_combined_pdf_for_attempt($assignment, $userid, $attemptnumber) {
-        global $CFG;
+    public static function get_combined_pdf_for_attempt($assignment, $userid, $attemptnumber) {
+        $document = self::get_combined_document_for_attempt($assignment, $userid, $attemptnumber);
 
-        require_once($CFG->libdir . '/pdflib.php');
-
-        $assignment = self::get_assignment_from_param($assignment);
-
-        if (!$assignment->can_view_submission($userid)) {
-            \print_error('nopermission');
-        }
-
-        $files = self::list_compatible_submission_files_for_attempt($assignment, $userid, $attemptnumber);
-
-        $pdf = new pdf();
-        if ($files) {
-            // Create a mega joined PDF.
-            $compatiblepdfs = array();
-            foreach ($files as $file) {
-                $compatiblepdf = pdf::ensure_pdf_compatible($file);
-                if ($compatiblepdf) {
-                    array_push($compatiblepdfs, $compatiblepdf);
-                }
-            }
-
-            $tmpdir = \make_temp_directory('assignfeedback_editpdf/combined/' . self::hash($assignment, $userid, $attemptnumber));
-            $tmpfile = $tmpdir . '/' . self::COMBINED_PDF_FILENAME;
-
-            @unlink($tmpfile);
-            try {
-                $pagecount = $pdf->combine_pdfs($compatiblepdfs, $tmpfile);
-            } catch (\Exception $e) {
-                debugging('TCPDF could not process the pdf files:' . $e->getMessage(), DEBUG_DEVELOPER);
-                // TCPDF does not recover from errors so we need to re-initialise the class.
-                $pagecount = 0;
-            }
-            if ($pagecount == 0) {
-                // We at least want a single blank page.
-                debugging('TCPDF did not produce a valid pdf:' . $tmpfile . '. Replacing with a blank pdf.', DEBUG_DEVELOPER);
-                @unlink($tmpfile);
-                $files = false;
-            }
-        }
-        $pdf->Close(); // No real need to close this pdf, because it has been saved by combine_pdfs(), but for clarity.
-
-        $grade = $assignment->get_user_grade($userid, true, $attemptnumber);
-        $record = new \stdClass();
-
-        $record->contextid = $assignment->get_context()->id;
-        $record->component = 'assignfeedback_editpdf';
-        $record->filearea = self::COMBINED_PDF_FILEAREA;
-        $record->itemid = $grade->id;
-        $record->filepath = '/';
-        $record->filename = self::COMBINED_PDF_FILENAME;
-        $fs = \get_file_storage();
-
-        $fs->delete_area_files($record->contextid, $record->component, $record->filearea, $record->itemid);
-
-        // Detect corrupt generated pdfs and replace with a blank one.
-        if ($files) {
-            $verifypdf = new pdf();
-            $pagecount = $verifypdf->load_pdf($tmpfile);
-            if ($pagecount <= 0) {
-                $files = false;
-            }
-            $verifypdf->Close(); // PDF loaded and never saved/outputted needs to be closed.
-        }
-
-        if (!$files) {
-            $file = $fs->create_file_from_string($record, base64_decode(self::BLANK_PDF_BASE64));
+        if ($document->get_status() === combined_document::STATUS_COMPLETE) {
+            // The combined document is already ready.
+            return $document;
         } else {
-            // This was a combined pdf.
-            $file = $fs->create_file_from_pathname($record, $tmpfile);
-            @unlink($tmpfile);
-
-            // Test the generated file for correctness.
-            $compatiblepdf = pdf::ensure_pdf_compatible($file);
+            // Attempt to combined the files in the document.
+            $grade = $assignment->get_user_grade($userid, true, $attemptnumber);
+            $document->combine_files($assignment->get_context()->id, $grade->id);
+            return $document;
         }
-
-        return $file;
     }
 
     /**
      * This function will return the number of pages of a pdf.
+     *
      * @param int|\assign $assignment
      * @param int $userid
      * @param int $attemptnumber (-1 means latest attempt)
@@ -377,27 +346,8 @@ EOD;
         }
 
         // Get a combined pdf file from all submitted pdf files.
-        $file = self::get_combined_pdf_for_attempt($assignment, $userid, $attemptnumber);
-        if (!$file) {
-            \print_error('Could not generate combined pdf.');
-        }
-
-        // Store the combined pdf file somewhere to be opened by tcpdf.
-        $tmpdir = \make_temp_directory('assignfeedback_editpdf/pagetotal/'
-            . self::hash($assignment, $userid, $attemptnumber));
-        $combined = $tmpdir . '/' . self::COMBINED_PDF_FILENAME;
-        $file->copy_content_to($combined); // Copy the file.
-
-        // Get the total number of pages.
-        $pdf = new pdf();
-        $pagecount = $pdf->set_pdf($combined);
-        $pdf->Close(); // PDF loaded and never saved/outputted needs to be closed.
-
-        // Delete temporary folders and files.
-        @unlink($combined);
-        @rmdir($tmpdir);
-
-        return $pagecount;
+        $document = self::get_combined_pdf_for_attempt($assignment, $userid, $attemptnumber);
+        return $document->get_page_count();
     }
 
     /**
@@ -407,7 +357,7 @@ EOD;
      * @param int $attemptnumber (-1 means latest attempt)
      * @return array(stored_file)
      */
-    public static function generate_page_images_for_attempt($assignment, $userid, $attemptnumber) {
+    protected static function generate_page_images_for_attempt($assignment, $userid, $attemptnumber) {
         global $CFG;
 
         require_once($CFG->libdir . '/pdflib.php');
@@ -419,14 +369,19 @@ EOD;
         }
 
         // Need to generate the page images - first get a combined pdf.
-        $file = self::get_combined_pdf_for_attempt($assignment, $userid, $attemptnumber);
-        if (!$file) {
-            throw new \moodle_exception('Could not generate combined pdf.');
+        $document = self::get_combined_pdf_for_attempt($assignment, $userid, $attemptnumber);
+
+        $status = $document->get_status();
+        if ($status === combined_document::STATUS_FAILED) {
+            print_error('Could not generate combined pdf.');
+        } else if ($status === combined_document::STATUS_PENDING_INPUT) {
+            // The conversion is still in progress.
+            return [];
         }
 
         $tmpdir = \make_temp_directory('assignfeedback_editpdf/pageimages/' . self::hash($assignment, $userid, $attemptnumber));
         $combined = $tmpdir . '/' . self::COMBINED_PDF_FILENAME;
-        $file->copy_content_to($combined); // Copy the file.
+        $document->get_combined_file()->copy_content_to($combined); // Copy the file.
 
         $pdf = new pdf();
 
@@ -641,10 +596,17 @@ EOD;
         }
 
         // Need to generate the page images - first get a combined pdf.
-        $file = self::get_combined_pdf_for_attempt($assignment, $userid, $attemptnumber);
-        if (!$file) {
-            throw new \moodle_exception('Could not generate combined pdf.');
+        $document = self::get_combined_pdf_for_attempt($assignment, $userid, $attemptnumber);
+
+        $status = $document->get_status();
+        if ($status === combined_document::STATUS_FAILED) {
+            print_error('Could not generate combined pdf.');
+        } else if ($status === combined_document::STATUS_PENDING_INPUT) {
+            // The conversion is still in progress.
+            return false;
         }
+
+        $file = $document->get_combined_file();
 
         $tmpdir = \make_temp_directory('assignfeedback_editpdf/final/' . self::hash($assignment, $userid, $attemptnumber));
         $combined = $tmpdir . '/' . self::COMBINED_PDF_FILENAME;
